% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler.R
\name{compile}
\alias{compile}
\title{Compiles a brainfuck program to a function}
\usage{
compile(brainfuck_code)
}
\arguments{
\item{brainfuck_code}{the brainfuck code}
}
\value{
It returns a function that expects a raw vectors which
is the internal memory of the brainfuck program.
According to Wikipedia you should pass in a raw vector of size 30,000 or larger.
}
\description{
Compiles a brainfuck program to a function
}
